Problema com o delete button quando o updatemode está on
O problema foi solucionado importando a variável updateMode para o código deleteItem.js e foi colocado uma restrição que só permite a execução do deleteButton se o updateMode estiver off

+++ js/deleteItem.js
@@ -1,11 +1,15 @@
 import { apiRequests } from "./requests.js";
 import { showItems } from "./showItems.js";
+import { updateMode } from "./updateItem.js";
 
 const url = "http://localhost:8080/items"
 
 async function deleteItem(event, item) {
     event.preventDefault();
-    
+    if(updateMode){
+        console.log('%cupdateItem.js line:51 Mensagem de erro: ', 'color: #007acc;', "You can't delete itens when update mode is on!");
+        return;
+    }
     await apiRequests.deleteItem(url, item.id);
 
     showItems.buildList();
